changes in storage

+++ src/main/java/com/jsp/warehouse_management_system/responsedto/WareHouseResponse.java
@@ -15,5 +15,5 @@ public class WareHouseResponse {
 
 	private int wareHouseId;
 	private String name;
-	private int totalCapacityInKg;
+	private double totalCapacityInKg;
 }
